Create cobot_pick_and_place.py
Note that this function assumes that the cobot arm is connected to the computer running the Python script, and that the joints of the arm are named joint1, joint2, joint3, etc. The function takes as input a list of joint names for the arm, the name of the object to be picked up, and the name of the joint to which the object should be dropped off. The function will move the arm to the pickup position, pick up the object, move the arm to the drop-off position, and drop off the object.